+++ extension/lib/bigcommerce/customer/jwt/TokenInvalidError.js
@@ -1,0 +1,11 @@
+class BigCommerceCustomerTokenInvalidError extends Error {
+  /**
+   * @param {string} token
+   */
+  constructor (token) {
+    super()
+    this.token = token
+  }
+}
+
+module.exports = BigCommerceCustomerTokenInvalidError