+++ extension/lib/shopgate/customer/autologin/InvalidTokenReceivedError.js
@@ -1,0 +1,9 @@
+class ShopgateAutologinInvalidTokenReceivedError extends Error {
+  constructor () {
+    super()
+    this.code = 'ELOGININVALIDTOKEN'
+    this.message = 'Provided token is not valid.'
+  }
+}
+
+module.exports = ShopgateAutologinInvalidTokenReceivedError