+++ extension/typedefs/BigCommerceCurrentCustomer.d.ts
@@ -1,0 +1,5 @@
+interface BigCommerceCurrentCustomer {
+  id: number
+  email: string
+  groupId: string
+}