+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- auto login pipeline for users to be able to login automatically after registering in BigCommerce store.
 
 ## [0.2.0] - 2018-04-17
 ### Added